Change the name of the Reactivity Handler and the API exposed
Change microTaskHandler to ReactivityManager
Change registerCallback to registerJob
Change waitNextTask to throttleRegisteredJob (next task was wrong, it was next microtask and + the name should not reflect the technology, just the expected result)
Add getReactivityProxy method to get a Proxy well configured for reactivity in this project
